Move budget templates to new directory
