+++ packages/desktop-client/src/components/budget/goals/editor/HistoricalAutomationReadOnly.tsx
@@ -3,7 +3,7 @@ import { Trans } from 'react-i18next';
 import type {
   CopyTemplate,
   AverageTemplate,
-} from 'loot-core/server/budget/types/templates';
+} from 'loot-core/types/models/templates';
 
 type HistoricalAutomationReadOnlyProps = {
   template: CopyTemplate | AverageTemplate;
