+++ packages/desktop-client/src/components/budget/goals/editor/PercentageAutomationReadOnly.tsx
@@ -1,6 +1,6 @@
 import { useTranslation, Trans } from 'react-i18next';
 
-import type { PercentageTemplate } from 'loot-core/server/budget/types/templates';
+import type { PercentageTemplate } from 'loot-core/types/models/templates';
 
 type PercentageAutomationReadOnlyProps = {
   template: PercentageTemplate;
