+++ packages/desktop-client/src/components/budget/goals/editor/ScheduleAutomationReadOnly.tsx
@@ -1,6 +1,6 @@
 import { Trans } from 'react-i18next';
 
-import type { ScheduleTemplate } from 'loot-core/server/budget/types/templates';
+import type { ScheduleTemplate } from 'loot-core/types/models/templates';
 
 type ScheduleAutomationReadOnlyProps = {
   template: ScheduleTemplate;
