+++ packages/desktop-client/src/components/budget/goals/editor/WeekAutomation.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from 'react-i18next';
 
-import type { WeekTemplate } from 'loot-core/server/budget/types/templates';
+import type { WeekTemplate } from 'loot-core/types/models/templates';
 
 import {
   type Action,
